use _fetch(); load only on execution
Two changes here, that I probably should have broken out.  The first is using
a before method modifier on run() to load most of our required modules via
class_load() when a consuming command's run() is called -- this should help
with startup time, particularly for the "reviewtool help" case.

The other command has to do with how we fetch files from remote locations.
URI::Fetch is very nice, but it has a tendency to silently and transparently
uncompress tarballs on the fly, so we prevent this from happening by poking at
a "private" variable to trick it into thinking that Compress::Zlib isn't
available.  For non-http requests, we use LWP::Simple (I was having problems
with ftp:// urls in specific with URI::Fetch).