Rework config; command cleanups
Our configfile is now just a flat .ini file; no more _sections() and the
mashing together of different parts of a .ini.  This should make it a
little... saner, I hope.

The commands were also retouched a bit with this; mainly style cleanups,
namespace::autoclean preferred over ::clean, POD touchups/snips and the moving
of module loading to run() whereever we could.